Refactored 'IMPROVED' Counter app files and directory names
Refactored files and directory names for 'IMPROVED' counter application to make more sense in the modularity of the application
