+++ U-C_TODO_frontend/src/course_learning/counter/IMPROVED_class_component_application/Counter.index.jsx
@@ -1,0 +1,3 @@
+import CounterApp from "./counter-component/CounterApp/CounterApp.component.jsx";
+
+export default CounterApp;
