+++ U-C_TODO_frontend/src/course_learning/counter/IMPROVED_class_component_application/counter-component/CounterApp/CounterApp.service.js
@@ -1,4 +1,4 @@
-import AppConstants from "./App.constant.js";
+import AppConstants from "./CounterApp.constant.js";
 
 export function resetButtonPress(resetButtonClickEventObj) {
   this.setState((previousState) => {
