+++ U-C_TODO_frontend/src/course_learning/counter/IMPROVED_class_component_application/counter-component/OperationPanel/OperationPanel.component.jsx
