--- U-C_TODO_frontend/src/course_learning/counter/IMPROVED_class_component_application/index.component.jsx
@@ -1,3 +1,0 @@
-import App from "./component/App/App.component.jsx";
-
-export default App;
